--- client/src/pages/ProfileListing/constants.js
@@ -1,62 +1,0 @@
-export default [
-  {
-    maintitle: 'Norma byers',
-    price: '$14/hr',
-    title: 'Loving pet sitter',
-    description: 'Dog sitting,cat sitting,pocket pet and bird core',
-    avatarUrl: 'https://upload.wikimedia.org/wikipedia/commons/e/e0/Keurig_Logo.png',
-    imageUrl:
-      'https://media.istockphoto.com/photos/studio-portrait-of-smiling-young-woman-holding-affectionate-pet-picture-id1170668075?k=20&m=1170668075&s=612x612&w=0&h=9BDN9UNU8D2o-dpLh_-Ow4SyCQy8L3B3eiz-F6JNpdo=',
-    country: 'Toronto,Ontario',
-  },
-  {
-    maintitle: 'Micheal Carnahan',
-    price: '$20/hr',
-    title: 'Loving pet sitter',
-    description: 'I would love to work with your dog',
-    avatarUrl: 'https://images-na.ssl-images-amazon.com/images/I/81CbxNLCHCL._AC_SL1500_.jpg',
-    imageUrl:
-      'https://media.istockphoto.com/photos/happy-young-beautiful-woman-hugging-her-favourite-pet-picture-id1181597935?k=20&m=1181597935&s=612x612&w=0&h=Pte0UIzAHA6B-BvLvAdt9b1WS4LEU469uz0k5qE0ezw=',
-    country: 'Toronto,Ontario',
-  },
-  {
-    maintitle: 'Jessica Pearson',
-    title: 'Dog care helper',
-    price: '$15/hr',
-    description: 'I have had dogs as pet for most of my life',
-    avatarUrl: 'https://images-na.ssl-images-amazon.com/images/I/81mjh5ZSu1L._AC_SL1500_.jpg',
-    imageUrl:
-      'https://media.istockphoto.com/photos/young-woman-and-her-cute-puppy-of-cocker-spaniel-outdoors-picture-id1204670115?k=20&m=1204670115&s=612x612&w=0&h=zynSgnSes68lb_VA6sjjvp-eULmF1f3mbW9s_RTc2uA=',
-    country: 'Toronto,Ontario',
-  },
-  {
-    maintitle: 'Charles Compton',
-    title: 'Passionate pet sitter',
-    price: '$15/hr',
-    description: 'I provide dog walking and pet itting services',
-    avatarUrl: 'https://images-na.ssl-images-amazon.com/images/I/61u76aiQQsL._AC_SL1000_.jpg',
-    imageUrl:
-      'https://media.istockphoto.com/photos/studio-portrait-of-smiling-young-woman-holding-little-dog-picture-id1311526307?k=20&m=1311526307&s=612x612&w=0&h=y-JLMkflmcJJb0pTy5xzFZ8pCLqCs5a3X7JvRVlMnKk=',
-    country: 'Toronto,Ontario',
-  },
-  {
-    maintitle: 'Charlotte Butler',
-    title: 'Animal Lover',
-    price: '$17/hr',
-    description: 'I am dog walker that cant wait to meet your pet!',
-    avatarUrl: 'https://images-na.ssl-images-amazon.com/images/I/81BdxpiRtCL._AC_SL1500_.jpg',
-    imageUrl:
-      'https://media.istockphoto.com/photos/woman-in-knitted-sweater-with-her-maltipoo-poodle-picture-id1266979016?k=20&m=1266979016&s=612x612&w=0&h=LuLW6lcVQuXeLWCBd_8PXZuflETnQqZo4E9poI1QIzg=',
-    country: 'Toronto,Ontario',
-  },
-  {
-    maintitle: 'Joan Blakeney',
-    title: 'Loving pet walker',
-    price: '$13/hr',
-    description: 'I have experience walking and working with dogs',
-    avatarUrl: 'https://images-na.ssl-images-amazon.com/images/I/61PGHBLkl8L._AC_SL1500_.jpg',
-    imageUrl:
-      'https://media.istockphoto.com/photos/happy-pretty-woman-embracing-yorkshire-terrier-picture-id923578330?k=20&m=923578330&s=612x612&w=0&h=c-YI7vajMInzXmVUcjY0Sg1fQhF1cXdb4t_YEiJeC8M=',
-    country: 'Toronto,Ontario',
-  },
-];
