+++ client/src/pages/ProfileListing/interface.ts
@@ -1,0 +1,9 @@
+export interface ProfileInterface {
+  avatarUrl: string;
+  title: string;
+  description: string;
+  imageUrl: string;
+  maintitle: string;
+  price: string;
+  country: string;
+}
